Organize project according to Golang standards: https://github.com/golang-standards/project-layout
