+++ README.md
@@ -1,2 +1,2 @@
-# parser
+# AsyncAPI Parser
 It parses AsyncAPI documents.
