--- schemaparser/avro.go
@@ -1,1 +1,0 @@
-package schemaparser
